+++ VictoryCenter/VictoryCenter.BLL/Commands/Admin/Partners/Create/CreatePartnersSectionCommand.cs
@@ -1,0 +1,7 @@
+﻿using FluentResults;
+using MediatR;
+using VictoryCenter.BLL.DTOs.Admin.Partners;
+
+namespace VictoryCenter.BLL.Commands.Admin.Partners.Create;
+
+public record CreatePartnersSectionCommand(CreatePartnersSectionDto CreatePartnersSectionDto) : IRequest<Result<PartnersSectionDto>>;
