+++ VictoryCenter/VictoryCenter.BLL/Commands/Admin/Partners/Delete/DeletePartnersSectionCommand.cs
@@ -1,0 +1,6 @@
+﻿using FluentResults;
+using MediatR;
+
+namespace VictoryCenter.BLL.Commands.Admin.Partners.Delete;
+
+public record DeletePartnersSectionCommand(long Id) : IRequest<Result<long>>;
