+++ VictoryCenter/VictoryCenter.BLL/Commands/Admin/Partners/Update/UpdatePartnersSectionCommand.cs
@@ -1,0 +1,8 @@
+﻿using FluentResults;
+using MediatR;
+using VictoryCenter.BLL.DTOs.Admin.Partners;
+
+namespace VictoryCenter.BLL.Commands.Admin.Partners.Update;
+
+public record UpdatePartnersSectionCommand(UpdatePartnersSectionDto UpdateDto, long Id)
+    : IRequest<Result<PartnersSectionDto>>;
