+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/Partners/CreatePartnerDto.cs
@@ -1,0 +1,7 @@
+﻿namespace VictoryCenter.BLL.DTOs.Admin.Partners;
+
+public record CreatePartnerDto
+{
+    public string Description { get; init; } = null!;
+    public CreatePartnerImageDto Image { get; init; } = null!;
+}
