+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/Partners/CreatePartnerImageDto.cs
@@ -1,0 +1,7 @@
+﻿namespace VictoryCenter.BLL.DTOs.Admin.Partners;
+
+public record CreatePartnerImageDto
+{
+    public string Base64 { get; init; } = null!;
+    public string MimeType { get; init; } = null!;
+}
