+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/Partners/PartnerDto.cs
@@ -1,0 +1,10 @@
+﻿using VictoryCenter.BLL.DTOs.Common;
+
+namespace VictoryCenter.BLL.DTOs.Admin.Partners;
+
+public record PartnerDto
+{
+    public long Id { get; init; }
+    public string Description { get; init; }
+    public ImageDto Image { get; init; } = null!;
+}
