+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/Partners/PartnersSectionDto.cs
@@ -1,0 +1,9 @@
+﻿namespace VictoryCenter.BLL.DTOs.Admin.Partners;
+
+public record PartnersSectionDto
+{
+    public long Id { get; init; }
+    public string Title { get; init; } = null!;
+    public string Description { get; init; } = null!;
+    public List<PartnerDto> Partners { get; init; } = [];
+}
