+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/Partners/UpdatePartnerImageDto.cs
@@ -1,0 +1,6 @@
+﻿namespace VictoryCenter.BLL.DTOs.Admin.Partners;
+
+public record UpdatePartnerImageDto : CreatePartnerImageDto
+{
+    public long? ImageId { get; init; } = null;
+}
