+++ VictoryCenter/VictoryCenter.BLL/Validators/Partners/CreatePartnerImageValidator.cs
@@ -1,0 +1,35 @@
+﻿using FluentValidation;
+using VictoryCenter.BLL.Constants;
+using VictoryCenter.BLL.DTOs.Admin.Images;
+using VictoryCenter.BLL.DTOs.Admin.Partners;
+
+namespace VictoryCenter.BLL.Validators.Partners;
+
+public class CreatePartnerImageValidator : AbstractValidator<CreatePartnerImageDto>
+{
+    public static readonly string[] AllowedMimeTypes = { "image/jpeg", "image/jpg", "image/png", "image/webp" };
+
+    public CreatePartnerImageValidator()
+    {
+        RuleFor(x => x.Base64)
+            .NotEmpty()
+            .WithMessage(ErrorMessagesConstants.PropertyIsRequired(nameof(UpdateImageDto.Base64)))
+            .Must(IsValidBase64).WithMessage(ImageConstants.Base64ValidationError);
+
+        RuleFor(x => x.MimeType)
+            .NotEmpty().WithMessage(ErrorMessagesConstants.PropertyIsRequired(nameof(UpdateImageDto.MimeType)))
+            .Must(mimeType => AllowedMimeTypes.Contains(mimeType))
+            .WithMessage(ImageConstants.MimeTypeValidationError(AllowedMimeTypes));
+    }
+
+    private static bool IsValidBase64(string? base64)
+    {
+        if (string.IsNullOrWhiteSpace(base64))
+        {
+            return false;
+        }
+
+        Span<byte> buffer = new(new byte[base64.Length]);
+        return Convert.TryFromBase64String(base64, buffer, out _);
+    }
+}
