+++ VictoryCenter/VictoryCenter.BLL/Validators/Partners/CreatePartnerValidator.cs
@@ -1,0 +1,23 @@
+﻿using FluentValidation;
+using VictoryCenter.BLL.Constants;
+using VictoryCenter.BLL.DTOs.Admin.Partners;
+
+namespace VictoryCenter.BLL.Validators.Partners;
+
+public class CreatePartnerValidator : AbstractValidator<CreatePartnerDto>
+{
+    public static readonly int DescriptionMaxLength = 50;
+
+    public CreatePartnerValidator()
+    {
+        RuleFor(x => x.Description)
+            .NotEmpty()
+            .WithMessage(ErrorMessagesConstants.PropertyIsRequired(nameof(CreatePartnerDto.Description)))
+            .MaximumLength(DescriptionMaxLength)
+            .WithMessage(ErrorMessagesConstants
+                .PropertyMustHaveAMaximumLengthOfNCharacters(nameof(CreatePartnerDto.Description), DescriptionMaxLength));
+
+        RuleFor(x => x.Image)
+            .SetValidator(new CreatePartnerImageValidator());
+    }
+}
