+++ VictoryCenter/VictoryCenter.DAL/Entities/Partner.cs
@@ -8,8 +8,8 @@ public class Partner : IOrderableEntity
     public long PartnersSectionId { get; set; }
     public string Description { get; set; } = null!;
     public long Priority { get; set; }
-    public DateTime CreatedAt { get; set; }
-    public long? ImageId { get; set; }
-    public Image? Image { get; set; }
+    public long ImageId { get; set; }
+    public DateTimeOffset CreatedAt { get; set; }
+    public Image Image { get; set; } = null!;
     public PartnerSection PartnerSection { get; set; } = null!;
 }
