+++ VictoryCenter/VictoryCenter.DAL/Entities/PartnerSection.cs
@@ -8,6 +8,6 @@ public class PartnerSection : IOrderableEntity
     public string Title { get; set; } = null!;
     public string Description { get; set; } = null!;
     public long Priority { get; set; }
-    public DateTime CreatedAt { get; set; }
+    public DateTimeOffset CreatedAt { get; set; }
     public ICollection<Partner> Partners { get; set; } = [];
 }
